+++ src/cli/lambdapi.ml
@@ -29,16 +29,23 @@ let sig_state_of_require =
     Core.Sig_state.of_sign
      (Compile.compile (Parsing.Parser.path_of_string req))
 
-let search_cmd cfg rules require s =
+(**[dbpath custom_dbpath] reads a string option [custom_dbpath] and returns
+  the path wrapped in it or a default value if None*)
+let dbpath custom_dbpath =
+  match custom_dbpath with
+  | None -> Common.Path.default_dbpath
+  | Some path -> path
+
+let search_cmd cfg rules require s custom_dbpath =
  Config.init cfg;
  let run () =
   Tool.Indexing.load_rewriting_rules rules ;
   let ss = sig_state_of_require require in
   out Format.std_formatter "%s@."
-   (Tool.Indexing.search_cmd_txt ss s) in
+   (Tool.Indexing.search_cmd_txt ss s (dbpath custom_dbpath)) in
  Error.handle_exceptions run
 
-let websearch_cmd cfg rules port require header_file =
+let websearch_cmd cfg rules port require header_file custom_dbpath =
  Config.init cfg;
  let run () =
   Tool.Indexing.load_rewriting_rules rules ;
@@ -81,10 +88,10 @@ let websearch_cmd cfg rules port require header_file =
     </ul>
       "
     | Some file -> Lplib.String.string_of_file file in
-  Tool.Websearch.start header ss ~port () in
+  Tool.Websearch.start header ss ~port (dbpath custom_dbpath) () in
  Error.handle_exceptions run
 
-let index_cmd cfg add_only rules files =
+let index_cmd cfg add_only rules files custom_dbpath=
  Config.init cfg;
  let run () =
   if not add_only then Tool.Indexing.empty ();
@@ -97,7 +104,7 @@ let index_cmd cfg add_only rules files =
    Tool.Indexing.load_rewriting_rules rules;
    Tool.Indexing.index_sign (no_wrn Compile.compile_file file) in
   List.iter handle files;
-  Tool.Indexing.dump () in
+  Tool.Indexing.dump (dbpath custom_dbpath) () in
  Error.handle_exceptions run
 
 end
@@ -466,6 +473,11 @@ let require_arg : string option CLT.t =
     "LP file to be required before starting the search engine." in
   Arg.(value & opt (some string) None & info ["require"] ~docv:"PATH" ~doc)
 
+let custom_db_path : string option CLT.t =
+  let doc =
+    "Path to the search DB file." in
+  Arg.(value & opt (some string) None & info ["db"] ~docv:"PATH" ~doc)
+
 let header_file_arg : string option CLT.t =
   let doc =
     "html file holding the header of the web page of the server." in
@@ -476,20 +488,20 @@ let index_cmd =
  let doc = "Index the given files." in
  Cmd.v (Cmd.info "index" ~doc ~man:man_pkg_file)
   Cmdliner.Term.(const LPSearchMain.index_cmd $ Config.full
-   $ add_only_arg $ rules_arg $ files)
+   $ add_only_arg $ rules_arg $ files $ custom_db_path)
 
 let search_cmd =
  let doc = "Run a search query against the index." in
  Cmd.v (Cmd.info "search" ~doc ~man:man_pkg_file)
   Cmdliner.Term.(const LPSearchMain.search_cmd $ Config.full
-   $ rules_arg $ require_arg $ query_as_arg)
+   $ rules_arg $ require_arg $ query_as_arg $ custom_db_path)
 
 let websearch_cmd =
  let doc =
   "Starts a webserver for searching the library." in
  Cmd.v (Cmd.info "websearch" ~doc ~man:man_pkg_file)
   Cmdliner.Term.(const LPSearchMain.websearch_cmd $ Config.full
-   $ rules_arg $ port_arg $ require_arg $ header_file_arg )
+   $ rules_arg $ port_arg $ require_arg $ header_file_arg $ custom_db_path)
 
 let _ =
   let t0 = Sys.time () in