+++ src/tool/websearch.mli
@@ -1,1 +1,2 @@
-val start : string -> Core.Sig_state.sig_state -> port:int -> unit -> unit
+val start : string -> Core.Sig_state.sig_state
+    -> port:int -> string -> unit -> unit