+++ CHANGES.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 - Add tactic `set`.
 - Decimal notation for integers.
+- add the `--db` option to the `index`, `search` and `websearch` commands
 
 ### Fixed